Supported for encrypted secrets with AES-GCM (#37)
This writes the secrets to an encrypted file on disk. A password is requested when using the secrets manager, and this is hashed and used as a key for encrypting/decrypting the contents. A new nonce is generated on each write to the file, and is stored alongside the encrypted data.

There are two issues which I would like to iterate on in future PRs:

1) Right now, you need to specify that you want the encrypted provider
each time you run vt with the --secrets-provider=encrypted flag. I
think we should add some kind of configuration file for vt, and allow
a default to be specified.

2) It is necessary to either enter the password on each execution, or specify
an environment variable with the password (VIBETOOL_SECRETS_PASSWORD).
I plan on investigating an integration with an OS keychain in future.