Define configuration file structure for Vibe Tool (#60)
The config file is a YAML file which is stored in an XDG-friendly
location. On application startup, the config file is created if it does
not already exist, and its location is printed out.

The Secrets provider option has been converted from a flag to a setting
in the configuration. A future PR will add a subcommand to set the
secret provider type so the user does not need to edit the config file
by hand.