Add support for VSCode Insider editor
This allows VT to add MCP servers into the new VSCode MCP config
structure. Currently this is only available in the Insider build of
VScode so we are targeting this instead of the regular VScode build.

The VScode config structure is slightly different to the format used by
the other clients we support. This PR makes a first attempt at a more
modular configuration editing structure, but I suspect I will end up
rewriting a lot of this code in a future PR since it turned out more
clunky than I had hoped.