removes autoremove on run
we want to make the removing of a container/mcp server explicitly driven
via `thv rm`. in this case we can properly clean up client configs.
otherwise when `thv stop` is run, the container gets removed but not the
client configs for them, leaving orphaned configs that users have to
clean up manually

Signed-off-by: ChrisJBurns <29541485+ChrisJBurns@users.noreply.github.com>