feat: adds formatted logging
If you wanted to pass in custom messages with variables into logging functions,
you had to use fmt.Sprintf in addition to the logging function.
This PR adds the formatted logging functions and switches all uses of logging
over to the new formatted logging functions so the calling of the functions
are nicer to use. In addition, this commit improves the logger tests
so that it uses the real logger functions so we can verify that the logs
are definitely real.

Signed-off-by: ChrisJBurns <29541485+ChrisJBurns@users.noreply.github.com>