Begin refactoring container lifecycle logic out of CLI
In order to allow the functionality of THV to be called from the CLI or
REST API, we need to move business logic out of the implementation of
the CLI. This PR is a first step which takes the rm, stop, list and
delete commands and moves them into a high-level "lifecycle" package.
Various refactoring has been carried out to remove duplicated code.

Work for future PRs:

1) Move run and restart here. This will need refactoring of the code for
   starting a THV container.
2) Unit tests for the functionality in this package.
3) Refactoring some logic out of this lifecycle package, for example,
   proxy and client config management functions.