Move runner protocol code into pkg
Since this logic will be reused by the API, move it inside the pkg
package.

One minor piece of refactoring: Get rid of the `logDebug` function - the
logger code sets the log level to debug if the `--debug` flag is passed
to the CLI, so this should not be necessary.