Refactor logger singleton
Previously the logger singleton was a global variable which every module
had to import. In the interests of preventing different parts of code
from directly changing the logger, and to slightly reduce the amount of
code needed for each log statement - make the logger singleton private,
and private global functions to access.

The interface for the logger is retained so that we can plug in new
logger backends without changing some of the logger-specific code in
ToolHive.