Merge commit from fork
* Do not persist secrets in the runconfig

Previously we expanded secrets into environment variables before writing
the run config into the state store. This led to the unfortunate
situation where the run config contained both references to secrets as
well as the raw secrets themselves. Since the state store is not
encrypted, it would be possible for someone with access to the user's
home folder to read the secrets without needing to know the decryption
password.

This code moves the point at which the secret expansion happens to after
we persist the run config, and right before we start the MCP server.

* Fix issue with detached processes

* Configure transport after writing config

* linting

* use fmt.Println instead of empty log line