Add initial API server to ToolHive
This PR adds a `serve` command along with scaffolding for the API. Two
basic endpoints are provided - a healthcheck, and a version endpoint.

The HTTP server is not yet integrated with logging - I think we need to
consider getting rid of our own logging interface and use something like
zerolog instead.