Improve version handling for local builds (#375)
Previously the application defaulted to a version of "dev" when building thv from source. Change the logic to manufacture a version string consisting of "build-" followed by the first eight characters of the commit SHA. If the commit SHA is unavailable, the version will be "build-unknown".

As part of this change, the update checking logic will print a more meaningful message instead of blindly telling the user that an update is available.