Add config + logic for loading remote registry
This adds the ability to load the MCP server
registry from a remote URL. The URL from which the
registry is loaded can be set in the config file.
By default, the registry URL is not set in the
config. In this case, the local registry will be
loaded for backwards compatibility.

Signed-off-by: Nick Terry <32914587+nick-terry@users.noreply.github.com>