Use consistent representation of a workload in the CLI and API (#667)
Prior to this PR, each CLI did its own bespoke presentation of workload state, and the API just serialized internal structures. Change the workload manager to return the new Workload domain model, populate it from container state, and alter both API and CLI to use it.

This PR has the effect of removing a lot of duplicate code from the CLI, and centralizing in one place. It also changes the data returned by the CLI list command, and will change the response structure of the Get/List workload API endpoints.

Fixes: #633