Make stop operation async, add stop all operation
This makes the stop operation asynchronous. The CLI calls it in a
blocking manner, the API will stop the container in the background and
return a 202 as soon as it has validated that the named container
exists.

As part of this change, add a stop all operation as requested for the UI
project. This logic is exposed both as a CLI command and API endpoint.