Move squid code into its own file (#766)
Also de-dupe the logic between egress and ingress. The existing code had two separate functions which were nearly identical.

Note that this is a first step in a series of changes I plan to make. The eventual end goal is to make the runtime interface purely container and image based, and move workload operations into a separate interface.