Move image pull/check/build out of runtime interface (#775)
Since these operations do not make sense in k8s, and since they are at a different level of abstraction to the other methods in the runtime interface, move them into a separate interface. We may have to expand image management into a more general container management interface, this is TBD.

A noop image manager implementation is provided for the parts of the code shared between k8s and the CLI which attempt to pull images. This can probably be avoided if we separate the k8s logic more clearly from the CLI workflow.