Refactor run command so logic is shared with API.
Previously, a lot of the logic relating to running a workload was only
implemented for the CLI. This PR addresses this so that most of the
logic from the run command now lives inside the constructor for the
RunConfig struct.

One of the more noteworthy challenges was to handle the code which
prompts users for missing env vars and secrets when the user attempts to
run an image from our trusted registry. For the API, k8s, and detached
processes, we do not want this to happen. To solve this problem I used
the strategy pattern - the CLI uses an implementation which prompts for
the missing secrets/env vars. The other use cases make use of an
implementation which throws an error if any of the required env vars or
secrets are missing.