Fix `no exposed ports specified in options.ExposedPorts` bug
The network isolation code expects a set of exposed ports to be
specified. The code which parses these ports was not included inside an
if block, meaning that it always was executed - even if network
isolation was disabled. Since the ports were not passed to this code, we
ended up with errors when running an MCP server. Specifically, the
container would be created, but the proxy runner process would die.

This PR fixes the bug when network isolation is disabled by ensuring
that this logic is only executed when the network isolation is disabled.
At this time, it's not clear where these values should come from when
network isolation is enabled - since they are not supplied by the `thv
run` command or API server. Any fixes needed will be in a follow up PR.