Ensure proxy process is killed during workload delete (#818)
Recently, DeleteWorkload was changed to not call the StopWorkload method. This was because the runtime will kill running containers for us. However, the call to delete the proxy runner process was deleted by mistake, leaving orphaned processes after workload deletion. This change will ensure the proxy runner process is cleaned up as part of deletion.