Use different logic for env vars depending on runtime
Prior to this PR, if ToolHive pulls a registry image and some of the
required env vars or secrets are missing, it will prompt the user to
enter them over stdin. This is not appropriate for the API or k8s
workflow. Place this logic behind an interface, and provide two
implementations - one which preserves the existing prompting behaviour,
and a second one which will error out if the variables are missing.

This PR is intended to further simplify the upcoming PR for sharing
workflow creation logic across the CLI and API.