Don't pull images in k8s (#889)
As part of refactoring the runtime interface, I caused the proxy process to attempt to pull down images when running in k8s. Handle this with a conditional guard for now. Given the current plan to split the k8s logic apart further, we may come up with a more natural split in future. At very least, we can consider having a no-op image puller.