Add authz config to the operator
Adds a new field `AuthzConfig` similar to `OIDCConfig` that can either
point to a configMap with authorization policy for a server or to an
inline policy.

The primary motivation is to be able to say that an agent can be
authorized to access this MCP server using its service account, e.g.:
```
authzConfig:
type: inline
inline:
    policies:
    - 'permit(principal == Client::"system:serviceaccount:mcop:search-summarizer-sa", action, resource);'
    entitiesJson: "[]"
```
Only allows the SA named `search-summarizer-sa` from the `mcop` namespace.

Alternatively refer to a configMap:
```
authzConfig:
type: configMap
configMap:
    name: search-summarizer-authz-external
    key: authz.json
```

The CM might look like:
```
apiVersion: v1
kind: ConfigMap
metadata:
  name: search-summarizer-authz-external
  namespace: mcop
data:
  authz.json: |
    {
      "policies": [
        "permit(principal == Client::\"system:serviceaccount:mcop:search-summarizer-sa\", action, resource);"
      ]
    }
```