Attach Network Isolation level to MCP server containers
There are some use cases which require us to know whether or not an MCP
workload was created with network isolation. This is solved by attaching
a label to those workload containers, and providing a method to query
whether the network isolation label was set.

As part of this change, I added some logic to the stop and remove workload
methods which check to see if this label is set before attempting to stop/delete
the proxy containers. This will get rid of spurious warnings when
stopping or deleting workloads where network isolation was disabled.