Create new StatusManager, move Get/List operations to this interface
Create the new StatusManager interface, which encapsulates the new state
tracking functionality. Since the operations to get details on workloads
is now dependent on the new state tracking, the Get/List workload
methods from the workload manager are moved here. Existing code has been
changed to call this new interface. To reduce the amount of change in
the PR, a simple implementation of state tracking is provided which
reuses the code for the old state tracking logic.

The creation of the StatusManager may seem like a strange split from the
existing workloads.Manager interface, but this is likely to be rectified
in future PRs - as part of the k8s work, and other changes, the existing
workloads.Manager is likely to be split further and renamed.

Fixes: #980