Attach Network Isolation label to MCP server containers (#967)
There are some use cases which require us to know whether or not an MCP workload was created with network isolation. This is solved by attaching a label to those workload containers, and providing a method to query whether the network isolation label was set.

As part of this change, I added some logic to the stop and remove workload methods which check to see if this label is set before attempting to stop/delete the proxy containers. However, I decided to comment it out since we need to consider that workloads created by older thv versions will not set the label. Once we have reasonable confidence that all workloads with network isolation have the right label set, we can reinstate these checks which will get rid of some unnecessary warnings that we see in thv when deleting/stopping workloads which do not have network isolation enabled.

Fixes: #961