Don't attempt to remove proxy containers when network isolation is off
This extends the work started in #967 by explicitly setting the network
isolation label to false when a new workload is created with network
isolation disabled. Workloads created before this change will not have
the label, and in order to ensure that it is cleaned up properly - we
treat those legacy workloads as having network isolation enabled. This
means that the problem described in #795 will continue, but those
workloads will be cleaned up over time and the problem will go away.