Don't attempt to remove proxy containers when network isolation is off (#994)
This extends the work started in #967 by explicitly setting the network isolation label to false when a new workload is created with network isolation disabled.

Workloads created before this change will not have the label, and in order to ensure that it is cleaned up properly - we treat all of those legacy workloads as having network isolation enabled. To minimize the impact of this, we log the proxy container not found errors at debug level.