Fix bug where the external network is killed during run/restart
When ToolHive starts or restarts a workload, it checks to see if a
container with the same name exists. If the configuration has changed,
it needs to remove the container and recreate it.

Previously, the code used the `RemoveWorkload` method to do this. However,
when the logic to create a separate external network was added, the
`RemoveWorkload` method was extended to remove the external network if no
workloads were left. This led to a situation where ToolHive checks to
see if the external network exists during start/restart, then checks to
see if there is an existing container, and ends up deleting the external
network _after_ checking to ensure it exists.

It is not straightforward to move the external network check due to the
sequence of operations needed to support the ingress/egress proxies
(even for workloads where network isolation is disabled). Instead, I
refactored the code to split the container deletion logic from
`RemoveWorkload` into a private method, and I call that method directly
in the `handleExistingContainer` method.