Use `proxy-port` instead of `port` in the CLI (#1022)
I find that having `port` and `target-port` flags is confusing, because it's not clear what exactly the `port` flag is for. Add a `proxy-port` flag, but leave `port` around for backwards compatibility. This is a cosmetic CLI-only change - it does not change the structure of the run configs which are stored to disk, or the API for the UI.