Move Get/List back into workloads.Manager
In a previous PR, I decided to move the operations to Get/List workloads
out of the main manager interface, and into the status tracker. After
working on wiring up the status tracker, I've come to realize that a
simpler approach is to keep the Get/List methods in the workload
Manager, and have the workload manager call the status manager. This PR
undoes the previous move. This has been lifted out of a bigger PR to
keep the diff small.