fix: remove allow_transport from network permissions
Currently we run with squid only when we ask for network isolation.
So this feature is not useful, as squid is only capable of filtering
by layer 7 transport. So transports like tcp/udp were not working
As it does not provide a helpful functionality, and can be confusing
to the user, just remove it. Users could only filter by host/ports,
and only when running with network isolation

Closes: #1032