Treat `--proxy-port` conflict as fatal error.
Prior to this change, when a `--proxy-port` was explicitly set by the
user, and the port was already in use, a new random port would be
selected instead. Since the common use case is to not select an explicit
proxy port, this behaviour led to some surprising results when a user
wanted to make the proxy listen on a port of their chosing.

Modify the code to error out when the user set the proxy port to a port
which is already in use. Note that this does not change the behaviour
when the proxy port is _not_ set - the port will still be randomly
selected in that case.

This PR includes some refactoring to use `proxy port` as the name for
this value instead of the old name (`port`).