Treat `--proxy-port` conflict as fatal error. (#1085)
Prior to this change, when --proxy-port was explicitly set by the user, and the port was already in use, a new random port would be selected instead. The common use case is to not set this flag and let thv pick a random port. Selecting a random port when the user actually requested a port explicitly has caused confusion for people who need thv to listen on a specific port.

Modify the code to error out when the user set the proxy port to a port which is already in use. Note that this does not change the behaviour when the proxy port is not set - the port will still be randomly selected in that case.

This PR includes some refactoring to use proxy port as the name for this value instead of the old name (port).