Split Runtime interface between Deployer and Runtime (#1060)
As we split apart the k8s proxy runner from the CLI/API code, we need to decouple the code in runner and transport from concerns which are not relevant to k8s.

As a first step, split the runtime interface in two: a Deployer interface which contains the subset of Runtime needed for the code in runner and transport, and leave the other methods in Runtime. Runtime is altered to extend from Deployer in order to minimize changes in the code, but in future we could decoupe them further to allow for implementations of Runtime if it makes sense to do so.