Use new MCP config path for VS Code 1.102 and later
Release 1.102 of VS Code expects MCP configuration to be defined in an
`mcp.json` file instead of the `settings.json` file. Change the
configuration code to use this new file instead.

As part of this change, we debated whether we should support the old
path as well. We decided not to since:

1) VS Code pushes updates aggressively, so we do not think that old
   versions will remain in use long term.
2) MCP support in the older releases was in a "preview" mode, so it was
   not intended to be stable long term.