Run client config CLI tests sequentially (#1092)
These are causing race conditions since they execute in parallel and need to set an environment variable in order to work.

Another issue I found: Now that we enforce that the proxy port is not already used, tests would fail locally when the sample port of 8080 was already in used.

Regarding the config env var race condition - In future, we may move the config to a database file, and we should address the lack of parameterization of the path at that point in time.