Add tool filtering to `MCPServer` CRD.
This change adds support for `tools` to `MCPServer` CRD.

By default, the list is `nil` which defaults to no `--tools` option
being passed as argument to the container. When a non-empty list is
passed, it is interpolated ensuring the items are sorted first, so
that changing the sequence of tools being enabled does not trigger a
reconciliation.

A simple unit test was added to verify the various cases of
reconciliation, but I also tested it locally using `kind` and a
debugger.

Fixes #1005