Stop returning ID outside of the layer, fix k8s ID bug
Currently a number of the runtime interface methods expect a container
ID to be passed. The format and meaning of this ID is specific to the
runtime. Callers of the runtime interface end up having to call the
runtime to get an ID and pass it back.

This changes all these methods to require a workload name instead of a
container ID. The runtime methods will do a lookup if needed to find the
container ID. This simplifies the logic in the workload manager.

While implementing this, I noticed a bug where the code gets mixed up
between pod IDs and stateful set names (which are the same as the name
of the workload). I changed the code to use the workload name to look up
stateful sets, and changed the code which operated on pods to search by
name and label instead of pod ID.