Use one workload status type consistently
Prior to this change, the Docker runtime returned Docker status values,
the kubernetes returned k8s statuses, and the workload manager attempted to
convert them to a single consistent enum type.

This code changes the runtime layer to return instances of our
WorkloadStatus enum, and removes the translation from the workload
manager.