Use one workload status type consistently (#1213)
Prior to this change, the Docker runtime returned Docker status values, the kubernetes code manufactured something which looked like Docker statuses, and the workload manager attempted to convert them to a single consistent enum type.

This code changes the runtime layer to return instances of our WorkloadStatus enum, and removes the translation from the workload manager.