Move GetWorkload/ListWorkload to status manager interface (#1222)
This moves the Get/List workload operations to the status manager interface. This is needed because the status manager implementation dictates which workloads can be found, and what their statuses should be.

The runtimeStatusManager implementation now relies on the runtime to provide details of which workloads are available, and their state. The fileStatusManager uses the status files, but also delegates to the runtime in order to handle workloads created before the status files were created.

Unit tests are added for workloads/status.go

Note that this does not wire in the file-based status tracking. That will be introduced in the next PR.