Fix bug where one workload name is a prefix substring of another
My recent change to use names as IDs included a function which uses the
Docker API to search for the matching workload container by name. The
Docker API uses prefix searching. If you have two workloads, and one
workload is a prefix substring of the other, searching for the shorter
name will match both and return an error (an error case I ironically
assumed would never happen...)

This changes the code to explicitly filter down the list using the exact
name in the case where more than one container is returned by Docker. I
have validated this by reproducing the scenario that was provided when
the bug was reported.