Support client configs with "." in json keys
Fix bug in `ensurePathExists()` in pkg/client/config_editor.go
where json keys containing "." would never be found, causing an
empty config path to always be returned in this case.

The gjson package treats "." as a special traversal character, so
when building the retrieval path, escape this character as `\.`
whenever it appears in json keys in the client config.

The hujson package treats "." as an ordinary character, so don't
apply any special handling when building the patch path.

Finally, unit testing revealed a bug where passing a path of "/"
(root path) causes hjson to attempt to always add a new key/value
pair `"": {}`. Fix this by handling the root path as a special
case, short circuiting all the traversal and patching logic.

Signed-off-by: Chris Gernon <17767786+while1fork@users.noreply.github.com>