Define types and structure for generic middleware creation
This is the first step towards generalizing the handling of middleware
in the code. This is intended to allow new middleware to be added
without altering the runner code or the runconfig schema.

This changes the runconfig to contain a list of generic MiddlewareConfig
structs, and modifies the runner to create middleware instances based on
this list, assuming a set of factory functions are supplied to the
runner.

Right now, this list is never populated, and none of the existing
middleware have been converted to use this scheme. This will be done in
a future PR.